testpynetworktables/requirements.txt
enum34
pynetworktables
